Change vm-mapping guarantees to be more performance friendly
While mapping the entire physical space up to the max address in the
memory map sounds like a convenient contract - it proves to be too
expensive performance-wise in cases where E820 contains very high memory
entries. A good example of this is the recent QEMU patch that introduced
a reserved E820 range at 0x000000FD00000000 -> 0x0000010000000000 for
AMD CPUs (which is set by default).

Change the contract to instead say that every memory range present in
the memory map is guaranteed to be covered by the page table, but
nothing on top of that. This should mitigate the issue without causing
much inconvenience.