Add support for i386 paging & a new page-table object
This extends the protocol to support most of the same features that we
had for amd64 kernels, but for the i386 targets.

It also adds a new page-table configuration object, which allows to
request a specific page-table layout as well as depth. null-guard is now
part of the the new page-table object.